refactor(!): improved type-narrowing when working with action settings (relates #133) (#164)
* refactor(!): improved type-narrowing when calling setSettings (relates #133)

* chore: address feedback

* test: fix types

* refactor(!): improve type-narrowing of getSettings in favour of casting, improve JSDocs.

* refactor: remove default settings type, and improve JSDocs

* refactor(!): further type-narrowing of settings

* fix: types and tests

* test: fix types

* docs: update changesets

* refactor: remove type obfuscation from getActionById

* chore: revert ignore

* test: fix types

* docs: grammar

* fix: action constructors

---------
