+++ .changeset/free-oranges-lose.md
@@ -1,0 +1,5 @@
+---
+"@elgato/streamdeck": major
+---
+
+`Action`, `KeyAction`, `DialAction`, and `NeoInfobarAction` types now require their setting type be defined.
