+++ .changeset/pink-jars-admire.md
@@ -1,0 +1,5 @@
+---
+"@elgato/streamdeck": major
+---
+
+Updated `getSettings` and `setSettings` so that the settings type is correctly bound to the type defined on the action.
