+++ .changeset/some-kings-taste.md
@@ -1,0 +1,5 @@
+---
+"@elgato/streamdeck": major
+---
+
+Updated `Action` type to represent a union of possible action types, such as key, dial, or Neo infobar.
