+++ packages/plugin/src/plugin/devices/device.ts
@@ -1,7 +1,7 @@
+import type { JsonObject } from "@elgato/utils";
+
 import type { DeviceInfo, DeviceType, Size } from "../../api/index.js";
-import type { DialAction } from "../actions/dial.js";
-import type { KeyAction } from "../actions/key.js";
-import type { NeoInfobarAction } from "../actions/neo-infobar.js";
+import type { Action } from "../actions/action.js";
 import { actionStore } from "../actions/store.js";
 import { connection } from "../connection.js";
 
@@ -62,7 +62,7 @@ export class Device {
 	 * Actions currently visible on the device.
 	 * @returns Collection of visible actions.
 	 */
-	public get actions(): IterableIterator<DialAction | KeyAction | NeoInfobarAction> {
+	public get actions(): IterableIterator<Action<JsonObject>> {
 		return actionStore.filter((a) => a.device.id === this.id);
 	}
